CCNs

Documentation for CCNs.

CCNs.CCN

CCNs.ccn

CCNs.infer_ccn_type

CCNs.CCN — Type

CCN

A representation of a CMS Certification Number.

CCNs are a uniform way of identifying providers or suppliers who currently or who ever have participated in the Medicare or Medicaid programs. A CCN is a 6- or 10-character alphanumeric string that encodes the provider or supplier's (respectively) State and facility type.

CCNs can be constructed from AbstractString or Integer objects, but Integers can only represent a subset of all possible CCNs.

CCNs are defined by CMS Manual System publication number 100-07 "State Operations Provider Certification".

CCNs.ccn — Method

ccn([T::Type], s)

Construct a CCN from input s.

Arguments

- `T::Type` (optional) - The type of the CCN. If no type is given, the best guess of the type will be made based on the format of the input `s`.
- `s::Union{AbstractString,Integer}` - The input to parse to create the CCN.

Return value

Returns a CCN of concrete type T if given, else the type will be inferred from format of s.

CCNs.infer_ccn_type — Method

infer_ccn_type(s) -> `Type{T} where T <: CCN`

Infer the type of the CCN from the input.

Arguments

- `s::Union{AbstractString,Integer}` - The value to parse.

Return value

The inferred type. Throws if the type cannot be inferred from the input.