CCNs

Documentation for CCNs.

CCNs.INVALID_FACILITY_TYPE

CCNs.INVALID_STATE

CCNs.CCN

CCNs.ccn

CCNs.decode

CCNs.facility_type

CCNs.facility_type_code

CCNs.infer_ccn_type

CCNs.sequence_number

CCNs.state

CCNs.state_code

CCNs.INVALID_FACILITY_TYPE — Constant

INVALID_FACILITY_TYPE

A String representing an invalid facility type code for a given CCN type.

CCNs.INVALID_STATE — Constant

INVALID_STATE

A String that represents an invalid state code.

CCNs.CCN — Type

CCN

A representation of a CMS Certification Number.

CCNs are a uniform way of identifying providers or suppliers who currently or who ever have participated in the Medicare or Medicaid programs. A CCN is a 6- or 10-character alphanumeric string that encodes the provider or supplier's (respectively) State and facility type.

CCNs can be constructed from AbstractString or Integer objects, but Integers can only represent a subset of all possible CCNs.

CCNs inherit from AbstractString, so methods like length, get, etc. are all defined and work as if the CCN were a string identifier.

CCNs are defined by CMS Manual System publication number 100-07 "State Operations Provider Certification".

CCNs.ccn — Method

ccn([T::Type], s)

Construct a CCN from input s.

Arguments

- `T::Type` (optional) - The type of the CCN. If no type is given, the best guess of the type will be made based on the format of the input `s`.
- `s::Union{AbstractString,Integer}` - The input to parse to create the CCN.

Return value

Returns a CCN of concrete type T if given, else the type will be inferred from format of s.

CCNs.decode — Function

decode([io::IO], ccn)

Decode ccn and either return the information as a String or write to IO.

CCNs.facility_type — Function

facility_type(ccn) -> String

Return a description of the facility type of ccn as a String.

Returns CCNs.INVALID_FACILITY_TYPE if the facility type code is invalid for the CCN type.

CCNs.facility_type_code — Function

facility_type_code(ccn) -> String

Return the facility type code of ccn as a String.

The facility type is dependent on the type of CCN, but usually involves the 3rd character of the code.

CCNs.infer_ccn_type — Method

infer_ccn_type(s) -> `Type{T} where T <: CCN`

Infer the type of the CCN from the input.

Arguments

- `s::Union{AbstractString,Integer}` - The value to parse.

Return value

The inferred type. Throws if the type cannot be inferred from the input.

CCNs.sequence_number — Function

sequence_number(ccn) -> Int64

Decode the sequence number from a given CCN.

Sequence numbers are sometimes indefinite. If this is the case, then only the decodable digits of the sequence number are returned (typically the last digits).

CCNs.state — Method

state(ccn) -> String

Decode the state code of ccn and return it as a String.

The first two characters of a CCN encode the "state" where the entity is located. "State" is interpreted loosely, as valid states include countries (like Canada) and territories (like (Guam).

Returns CCNs.INVALID_STATE if the first two characters are not a valid state code.

CCNs.state_code — Method

state_code(ccn) -> String

Return the state code of ccn (the first two characters) as a String.